+++ sketch-animated/src/commonTest/kotlin/com/github/panpf/sketch/animated/common/test/request/AnimatedExtensionsTest.kt
@@ -3,6 +3,7 @@ package com.github.panpf.sketch.animated.common.test.request
 import com.github.panpf.sketch.images.ResourceImages
 import com.github.panpf.sketch.request.ImageOptions
 import com.github.panpf.sketch.request.ImageRequest
+import com.github.panpf.sketch.request.animatedTransformation
 import com.github.panpf.sketch.request.animationEndCallback
 import com.github.panpf.sketch.request.animationStartCallback
 import com.github.panpf.sketch.request.onAnimationEnd
@@ -10,6 +11,8 @@ import com.github.panpf.sketch.request.onAnimationStart
 import com.github.panpf.sketch.request.repeatCount
 import com.github.panpf.sketch.test.singleton.getTestContextAndSketch
 import com.github.panpf.sketch.test.utils.toRequestContext
+import com.github.panpf.sketch.transform.AnimatedTransformation
+import com.github.panpf.sketch.transform.PixelOpacity
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -174,5 +177,59 @@ class AnimatedExtensionsTest {
             onAnimationEnd(myAnimationEndCallback)
         }.toRequestContext(sketch).cacheKey
         assertEquals(cacheKey1, cacheKey2)
+    }
+
+    @Test
+    fun testAnimatedTransformation() = runTest {
+        val (context, sketch) = getTestContextAndSketch()
+        val myAnimatedTransformation = TranslucentAnimatedTransformation
+
+        ImageRequest(context, ResourceImages.animGif.uri).apply {
+            assertNull(animatedTransformation)
+        }
+        ImageRequest(context, ResourceImages.animGif.uri) {
+            this.animatedTransformation(myAnimatedTransformation)
+        }.apply {
+            assertEquals(myAnimatedTransformation, animatedTransformation)
+        }
+
+        ImageRequest(context, ResourceImages.animGif.uri).apply {
+            assertNull(animatedTransformation)
+        }
+        ImageRequest(context, ResourceImages.animGif.uri) {
+            animatedTransformation(myAnimatedTransformation)
+        }.apply {
+            assertEquals(myAnimatedTransformation, animatedTransformation)
+        }
+
+        ImageOptions().apply {
+            assertNull(animatedTransformation)
+        }
+        ImageOptions {
+            animatedTransformation(myAnimatedTransformation)
+        }.apply {
+            assertEquals(myAnimatedTransformation, animatedTransformation)
+        }
+
+        val key1 = ImageRequest(context, ResourceImages.animGif.uri).key
+        val key2 = ImageRequest(context, ResourceImages.animGif.uri) {
+            animatedTransformation(myAnimatedTransformation)
+        }.key
+        assertNotEquals(key1, key2)
+
+        val cacheKey1 =
+            ImageRequest(context, ResourceImages.animGif.uri).toRequestContext(sketch).cacheKey
+        val cacheKey2 = ImageRequest(context, ResourceImages.animGif.uri) {
+            animatedTransformation(myAnimatedTransformation)
+        }.toRequestContext(sketch).cacheKey
+        assertEquals(cacheKey1, cacheKey2)
+    }
+
+    private data object TranslucentAnimatedTransformation : AnimatedTransformation {
+        override val key: String = "TranslucentAnimatedTransformation"
+
+        override fun transform(canvas: Any): PixelOpacity {
+            return PixelOpacity.TRANSLUCENT
+        }
     }
 }